refactor: moved types to single index.ts; cleaned up api/selected route
