+++ src/components/common/Icons/EditIcon.tsx
@@ -1,4 +1,4 @@
-import { SVGProps } from "@/types/SVGProps";
+import { SVGProps } from "@/types";
 
 export function EditIcon(props?: SVGProps) {
   return (
