+++ src/components/common/Icons/LogoIcon.tsx
@@ -1,4 +1,6 @@
-export function LogoIcon() {
+import { SVGProps } from "@/types";
+
+export function LogoIcon(props?: SVGProps) {
   return (
     <svg
       width="93"
@@ -6,6 +8,7 @@ export function LogoIcon() {
       viewBox="0 0 93 45"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
+      {...props}
     >
       <path
         d="M3.89234 27.3907C4.15846 27.3408 4.45785 27.3076 4.7905 27.3076C4.8404 25.5611 5.83835 22.6671 6.90284 19.6732C5.90488 19.5235 5.05662 19.5235 4.54101 19.8728C4.17509 18.4257 5.5556 18.143 7.45171 18.1763C8.43304 15.3986 9.3811 12.6875 9.58069 10.8745C8.00059 10.7747 6.60345 10.9411 5.72192 11.939C5.43917 12.2384 4.67407 8.46282 8.13365 7.83078C10.9446 7.31517 16.9323 9.61047 18.8118 6.96588C19.0945 6.58333 19.3274 10.0429 17.4479 10.9743C16.2504 11.5565 14.6703 11.49 13.0236 11.2904C12.7908 12.8871 11.9259 15.5317 11.0277 18.4091C12.9904 18.5422 14.7202 18.5255 15.0861 17.5941C15.6682 21.0703 13.19 20.9539 10.429 20.3884C9.66385 22.9498 8.94865 25.5445 8.68253 27.6568C11.0943 27.9396 13.506 28.0394 14.5705 26.5258C14.8532 26.1433 15.0861 29.6029 13.2066 30.5177C9.91334 32.1476 3.84244 28.8211 1.48061 31.4823C1.19785 31.7817 0.432754 28.0061 3.89234 27.3907ZM17.9136 17.1617C17.897 17.0951 21.5894 16.7625 21.5229 18.8915C21.473 20.2553 18.2795 27.241 18.9448 29.4864C20.5748 27.4074 24.916 17.1617 24.916 17.1617C24.8993 17.0951 28.5918 16.7625 28.5086 18.8915C28.4587 20.272 25.6644 25.8439 25.9472 29.5862C27.2279 28.239 28.7747 24.5299 29.0575 23.8313C29.0907 23.7648 29.1573 23.7316 29.2238 23.7316C29.3569 23.7316 29.44 23.848 29.4068 23.9644C29.0076 25.1952 27.3443 30.2682 26.4794 30.7838C25.4815 31.3992 22.7703 30.368 22.5209 29.2369C22.0385 27.1412 24.0677 22.1515 24.7496 19.141C23.419 22.6172 20.342 30.4179 19.2775 30.8836C18.2962 31.2994 15.635 30.4511 15.4187 29.1205C15.0196 26.7088 18.4292 19.0911 17.9136 17.1617ZM25.3151 38.9171C24.6332 34.7257 30.9869 20.6213 30.3215 17.1284C30.3049 17.0619 34.147 17.1284 33.9142 19.2407C33.698 21.2533 30.8372 29.187 29.4899 35.0916C31.9016 26.1433 36.5421 17.6773 37.2573 17.2615C37.906 16.8789 39.9019 17.4444 40.7169 18.8748C41.8147 20.8042 39.4861 30.1185 35.6273 30.9002C34.7625 31.0832 32.7166 29.8025 32.3008 29.2868C30.8704 32.9626 29.5398 37.0709 28.9078 40.8631C28.8246 41.2623 25.5979 40.6469 25.3151 38.9171ZM37.6066 18.3093C37.3904 18.2262 34.829 22.8667 32.4838 28.8045C35.5608 30.5343 38.3884 18.6253 37.6066 18.3093ZM40.3843 28.9875C39.7023 24.7794 46.1392 10.1926 45.7067 7.08231C45.6901 7.01578 49.2162 7.19874 48.9833 9.31108C48.7505 11.49 45.3741 20.6712 44.243 26.6589C45.8564 22.6671 48.6673 16.9621 50.2807 17.1118C51.2121 17.1949 53.0916 18.0765 53.2579 18.642C53.7236 20.2054 50.4636 27.3408 50.3971 29.7027C51.6944 28.4552 53.3244 24.5465 53.6072 23.8313C53.6405 23.7648 53.707 23.7316 53.7735 23.7316C53.9066 23.7316 53.9897 23.848 53.9565 23.9644C53.5573 25.1952 51.7277 30.8337 50.9959 30.9335C49.5488 31.1331 47.1038 30.0852 46.9043 28.7712C46.6548 26.9915 49.4989 21.22 49.7817 18.4257C48.9501 18.2927 43.9437 28.3055 43.9769 30.8836C44.0767 31.3493 40.6837 30.7505 40.3843 28.9875ZM53.4076 28.7546C52.4263 25.977 54.6551 19.0245 57.1666 17.4112C58.0315 16.8457 60.3767 16.9953 61.1251 17.4943C61.7738 17.9101 61.358 18.3093 61.358 18.3093C58.6469 17.0452 55.4534 28.8211 56.7009 29.9688C58.1479 31.2994 61.5909 20.405 59.5118 19.9726C59.5118 19.9726 60.4765 18.6087 61.3081 18.7085C64.1855 19.0911 62.0067 28.871 57.9816 30.7671C56.8838 31.2828 54.2725 31.183 53.4076 28.7546ZM62.1231 29.1538C61.6906 26.5092 64.9673 19.0578 64.435 17.1284C64.4184 17.0619 68.3104 16.7625 68.2439 18.8915C68.194 20.089 66.0817 24.4468 65.7157 28.0228C66.9798 24.0975 69.8739 16.9454 71.6702 17.1118C72.6016 17.1949 74.2316 17.96 74.398 18.5089C74.6142 19.2407 72.3854 22.7336 72.9842 24.8958C68.0277 24.6796 71.2378 19.0744 71.1712 18.4257C70.3396 18.2927 65.8821 28.2057 65.8987 30.8004C65.9153 30.9002 65.7989 30.9667 65.7157 30.9501C65.0005 30.8669 62.3393 30.4844 62.1231 29.1538ZM79.3545 16.8623C74.5975 16.2302 77.6247 13.5524 77.4251 11.7893C82.5812 12.621 78.9387 14.4007 79.3545 16.8623ZM72.9676 29.1538C72.5018 26.609 75.9282 19.0578 75.4125 17.1284C75.3959 17.0619 79.2547 16.7625 79.1881 18.8915C79.1383 20.272 76.3772 25.7607 76.6101 29.5031C77.8742 28.0893 79.3711 24.5133 79.6539 23.8313C79.6705 23.7648 79.737 23.7316 79.8202 23.7316C79.9366 23.7316 80.0198 23.848 79.9865 23.9644C79.604 25.1952 77.791 30.5177 76.9926 30.9002C76.1776 31.2661 73.1838 30.4012 72.9676 29.1538ZM82.9139 30.8836C82.3151 31.2994 79.6206 30.5676 79.0052 29.4532C78.2068 28.0228 80.2027 20.7709 82.0323 18.4257C83.4627 16.6128 86.5564 17.0619 87.2882 17.4943C87.4712 17.5941 87.5377 17.6939 87.4379 17.9767C87.3215 18.2927 87.0886 18.2262 86.7227 18.1264C84.8599 17.644 82.0656 27.8564 82.2486 29.9355L88.1864 17.2947C88.386 16.7459 91.9121 17.6606 91.5794 18.3592C91.2468 19.0578 88.8184 27.3408 88.7519 29.7027C90.0492 28.4552 91.6959 24.5465 91.9786 23.8313C92.0119 23.7648 92.0784 23.7316 92.145 23.7316C92.2614 23.7316 92.3612 23.848 92.3113 23.9644C91.9287 25.1952 90.0991 30.8337 89.3673 30.9335C87.9036 31.1331 85.4586 30.0852 85.2757 28.7712C85.1094 27.6568 86.0242 24.2139 86.8724 21.3864C85.4087 25.1786 83.2465 30.6507 82.9139 30.8836Z"
