+++ src/components/common/Icons/UserIcon.tsx
@@ -1,4 +1,6 @@
-export function UserIcon() {
+import { SVGProps } from "@/types";
+
+export function UserIcon(props?: SVGProps) {
   return (
     <svg
       width="20"
@@ -6,6 +8,7 @@ export function UserIcon() {
       viewBox="0 0 20 20"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
+      {...props}
     >
       <path
         d="M9.99992 11.6667C12.3011 11.6667 14.1666 9.8012 14.1666 7.50001C14.1666 5.19882 12.3011 3.33334 9.99992 3.33334C7.69873 3.33334 5.83325 5.19882 5.83325 7.50001C5.83325 9.8012 7.69873 11.6667 9.99992 11.6667ZM9.99992 11.6667C6.31802 11.6667 3.33325 13.9053 3.33325 16.6667M9.99992 11.6667C13.6818 11.6667 16.6666 13.9053 16.6666 16.6667"
