--- src/types/ButtonProps.ts
@@ -1,7 +1,0 @@
-import { ComponentPropsWithoutRef } from "react";
-
-type CustomProps = {
-  isSubmit?: boolean;
-};
-
-export type ButtonProps = ComponentPropsWithoutRef<"button"> & CustomProps;
