--- src/types/SVGProps.ts
@@ -1,3 +1,0 @@
-import React from "react";
-
-export type SVGProps = React.SVGProps<SVGSVGElement>;
