+++ src/types/index.ts
@@ -1,3 +1,5 @@
+import React, { ComponentPropsWithoutRef } from "react";
+
 export type SelectedItem = {
   id: string;
   isInWishlist?: boolean;
@@ -47,3 +49,11 @@ export type ProtectedRequest = {
     email: string;
   };
 } & Request;
+
+type CustomButtonProps = {
+  isSubmit?: boolean;
+};
+
+export type SVGProps = React.SVGProps<SVGSVGElement>;
+export type ButtonProps = ComponentPropsWithoutRef<"button"> &
+  CustomButtonProps;
